pyproject.toml
[project]
name = "cpk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "gitpython>=3.1.44",
    "toml>=0.10.2",
]

[project.scripts]
cpk = "cpk.__main__:cli"



